+++ test/srtm/source/esa_test.exs
@@ -9,4 +9,11 @@ defmodule SRTM.Source.ESATest do
     assert {:ok, <<5, 163, 5, 163, 5, 162, 5, 161, 5, 161>> <> _} =
              ESA.fetch("N36W117", endpoint: "http://localhost:#{bypass.port}/")
   end
+
+  test "returns an error if the response isn't a zip archive", %{bypass: bypass} do
+    expect_hgt_download(bypass, {200, "<html>captive portal</html>"})
+
+    assert {:error, %SRTM.Error{reason: :invalid_archive}} =
+             ESA.fetch("N36W117", endpoint: "http://localhost:#{bypass.port}/")
+  end
 end